// Sensalon_back/src/interfaces/Distributors.ts
export interface Distributor {
    iIdDistributor: string;
    iFIdUser?: string;
    vccompanies?: string;
    vcaddress?: string;
    vcstate?: string;
    vccity?: string;
    vcpostalcode?: string;
    vccountry?: string;
    vcphone?: string;
    vcemail?: string;
    vcrfc?: string;
    vcconsfis?: string; 
    dtcreation?: Date;
    dtupdate?: Date;
  }
  